refactor(wire): extract nested projection engine
## Summary

- Move the recursive nested and decorator projection engine into a
  cohesive `_wire_nested.py` module.
- Reduce `_wire.py` to four projection coordinators and direct
  compatibility re-exports.

## Boundaries and compatibility

- Preserve per-family compilation state, recursive placeholder caching,
  hash-required set projections, generated names, defaults, and error
  precedence.
- Preserve automatic omission for `exclude=True` and `exclude_if` fields
  across current, historical, nested, and decorator-discovered wire
  projections.
- Keep registry imports local and the runtime import graph acyclic.
  No public API or wire behavior changes.

## Investigation

- Partition all 28 original definitions exactly: four remain in the
  facade and `_WireCompilationContext` plus 23 functions move unchanged.
- Confirm moved and retained definitions are source- and AST-identical,
  and facade re-exports retain object identity.
- Track the parent architecture work in #95.

## Validation

- `uv run make ci`: 859 tests passed with 95.89% coverage.
- Pydantic 2.12.3 lower-bound run: 859 tests passed.
- `v0_3_0_contract.py --check`: passed against immutable 0.3.0 fixtures.
- Two independent read-only reviews found no blockers.

Refs #95, #122
